refactor userhouseholds to multiple selectors and place in slice
